* 2016. 2. 18. 컴파일오류  수정.
TODO.
계속 수정.

+++ Cabochon/CabochonProject/dx9_game_engine/image.h
@@ -20,8 +20,8 @@ protected:
 	int startFrame;
 	int endFrame;
 	int currentFrame;
-	double frameDelay;
-	double animTimer;
+	float frameDelay;
+	float animTimer;
 
 	HRESULT hr;
 	bool loop;
